started building out basic modal to add students. Remove students button not currently functional
